RunKOfams in kegg.py -> metabolism/annotate.py
